Add curve namespace.
The curve namespace defines the Curve protocol, which is basically just
a particular case of clojure.lang.IFn (since, curves are, after all,
literally functions in the purest mathematical sense), and the Linear
curve deftype. It does what you'd expect, and extends clojure.lang.IFn
as well (I wish the deftype didn't have to do that).

There is also a nicely named equivalent to the Linear constructor, named
`linear` (note the case change, and that no dot is needed).