Add util namespace, with `interpolate`.
The interpolate function is a clamped linear function between two
values. It relies on the absolute value function, `util/abs`.

The test of the interpolate function relies on a `mostly=` function,
which returns true when two numbers are less than a billionth apart.
This is generally useful, so add it to the proper util namespace, and
test it.