Add animation namespace.
Animations are simply curves with a defined start time and duration.
Their values are clamped to the values of its underlying curve at
t = 0 and t = 1.

Animations are created with the `animation` construction function. The
`animation` function can take keywords representing integers with units
like `:10s`. If two units are supplied and they're different, an error
will be thrown.

Animations should primarily be used as functions of time. However, they
are records, so they can also be used as maps.