Make animations & curves records, not deftypes.
This allows them to be introspected (for example, to get the control
points out of a bezier curve). Fortunately, it seems that the default
IFn implementation for records can be overridden, so their interface
does not change.

Add a clojure.lang.IFn/applyTo implementation for animations, and add
explicit tests that apply works with both animations and curves.